requirements.txt
tensorflow
pillow
tkcolorpicker
